Move GenServer capabilities from PhoenixSignaling into custom Registry module
