+++ lib/membrane_webrtc/app.ex
@@ -6,7 +6,7 @@ defmodule Membrane.WebRTC.App do
   def start(_opts, _args) do
     children =
       [{Registry, name: Membrane.WebRTC.WhipRegistry, keys: :unique}] ++
-        if Code.ensure_loaded?(Phoenix), do: [Membrane.WebRTC.PhoenixSignaling], else: []
+        if Code.ensure_loaded?(Phoenix), do: [Membrane.WebRTC.PhoenixSignaling.Registry], else: []
 
     Supervisor.start_link(children, strategy: :one_for_all, name: __MODULE__.Supervisor)
   end
