+++ lib/membrane_webrtc/phoenix_signaling.ex
@@ -3,7 +3,6 @@ if Code.ensure_loaded?(Phoenix) do
     @moduledoc """
     Provides signaling capabilities for WebRTC connections through Phoenix channels.
     """
-    use GenServer
     alias Membrane.WebRTC.Signaling
 
     @typedoc """
@@ -11,39 +10,6 @@ if Code.ensure_loaded?(Phoenix) do
     instances.
     """
     @type signaling_id :: String.t()
-
-    @spec start(term()) :: GenServer.on_start()
-    def start(args) do
-      GenServer.start(__MODULE__, args, name: __MODULE__)
-    end
-
-    @spec start_link(term()) :: GenServer.on_start()
-    def start_link(args) do
-      GenServer.start_link(__MODULE__, args, name: __MODULE__)
-    end
-
-    @impl true
-    def init(_args) do
-      {:ok, %{signaling_map: %{}}}
-    end
-
-    @impl true
-    def handle_call({:get_or_create, signaling_id}, _from, state) do
-      case Map.get(state.signaling_map, signaling_id) do
-        nil ->
-          signaling = Signaling.new()
-          state = put_in(state, [:signaling_map, signaling_id], signaling)
-          {:reply, signaling, state}
-
-        signaling ->
-          {:reply, signaling, state}
-      end
-    end
-
-    @impl true
-    def handle_call({:get, signaling_id}, _from, state) do
-      {:reply, Map.get(state.signaling_map, signaling_id), state}
-    end
 
     @doc """
     Returns an instance of a Phoenix Signaling associated with given signaling ID.
@@ -74,11 +40,11 @@ if Code.ensure_loaded?(Phoenix) do
     end
 
     defp get_or_create(signaling_id) do
-      GenServer.call(__MODULE__, {:get_or_create, signaling_id})
+      GenServer.call(__MODULE__.Registry, {:get_or_create, signaling_id})
     end
 
     defp get!(signaling_id) do
-      case GenServer.call(__MODULE__, {:get, signaling_id}) do
+      case GenServer.call(__MODULE__.Registry, {:get, signaling_id}) do
         nil ->
           raise "Couldn't find signaling instance associated with signaling_id: #{inspect(signaling_id)}"
